requirements.txt
requests
python-dotenv
openai
pydantic
fastapi
uvicorn
pytest
pytest-asyncio